Mambaru (#35)
* -WJSON_ENABLE_LONG_LONG, wjson.hpp include strerror
* Deserialization of flags as arrays for any delimiters
* Deserialization of flags as string for any delimiters
* flags deser init bugfix
* ex warn fix
* nullptr
* wrn fix g++-8
* nullptr for c++98

+++ wjson/serializer/array.hpp
@@ -191,7 +191,7 @@ public:
 };
 #endif // __cplusplus >= 201103L
 
-template< typename J, int N, typename RR, char L, char R >
+template< typename J, size_t N, typename RR, char L, char R >
 class serializerA< array_r< J[N], RR>, L, R >
 {
   typedef array_r< J[N], RR> array_type;
@@ -207,7 +207,7 @@ public:
   {
     if ( parser::is_null(beg, end) )
     {
-      for (int i = 0; i < N ; ++i)
+      for (size_t i = 0; i < N ; ++i)
       {
         // null для каждого элемента
         serializer()(t[i], beg, end, e);
